separate viewset classes from other function-based views for cleaner code organization
